update: tokenizer unit tests

+++ src/parsemath/token.rs
@@ -1,5 +1,7 @@
+#[derive(Debug)]
+#[derive(PartialEq)]
 pub enum Token {
-    Num(i64), // for each number token
+    Num(f64), // for each number token
     Add,      // for the add operator token
     Subtract,
     Multiply,
